+++ engine/src/transformations.rs
@@ -1,21 +1,21 @@
 //! Defines a trait that serves as a wrapper around noise modules and transform their inputs in some way.
 
-use noise::{NoiseModule, Point3};
+use noise::{NoiseFn, Point3};
 
 use composition_tree::CompositionTreeNode;
 use util::Dim;
 
 pub enum InputTransformation {
     /// Transforms the input coordinate by scaling the X and Y dims by `zoom` and the Z dim by `speed`.
-    ZoomScale {speed: f32, zoom: f32},
+    ZoomScale {speed: f64, zoom: f64},
     /// Replaces the `replacement_dim` of the input with the result of passing it through the provided noise function.
     HigherOrderNoiseModule {node: CompositionTreeNode, replaced_dim: Dim},
     /// Multiplies all of the input values by this value
-    ScaleAll(f32),
+    ScaleAll(f64),
 }
 
 impl InputTransformation {
-    pub fn transform(&self, coord: Point3<f32>) -> Point3<f32> {
+    pub fn transform(&self, coord: Point3<f64>) -> Point3<f64> {
         match self {
             &InputTransformation::ZoomScale{ speed, zoom } => {
                 [coord[0] * zoom, coord[1] * zoom, coord[2] * speed]
